# quwoquan_service/services/product-ops-service/generated/openapi.yaml
...
openapi: 3.0.3
info:
    title: quwoquan ops API
    version: canonical
paths:
...
    /control-plane/product/approvals:
        get:
            operationId: ListProductApprovals
            x-contract-operation-id: ops.event_record.ListProductApprovals
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordControlPlaneQueryFacade
                method: listProductApprovals
                reader: ProductApprovalReader
                slice: ProductApprovalSlice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListProductApprovalsResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /control-plane/product/audits:
        get:
            operationId: ListProductAudits
            x-contract-operation-id: ops.event_record.ListProductAudits
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordControlPlaneQueryFacade
                method: listProductAudits
                reader: ProductAuditReader
                slice: ProductAuditSlice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListProductAuditsResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /control-plane/product/experience/pages:
        get:
            operationId: GetPageExperience
            x-contract-operation-id: ops.event_record.GetPageExperience
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordMetricQueryFacade
                method: getPageExperience
                reader: PageExperienceReader
                slice: PageExperienceSlice
            parameters:
                - name: from
                  in: query
                  required: false
                  schema:
                    type: string
                - name: to
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetPageExperienceResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /control-plane/product/experiments:
        get:
            operationId: ListExperiments
            x-contract-operation-id: ops.experiment.ListExperiments
            x-object-id: ops.experiment
            x-actor: account
            x-application:
                kind: query
                facet: ExperimentQueryFacade
                method: listExperiments
                reader: ExperimentCatalogReader
                slice: ExperimentCatalogSlice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ExperimentCatalogSlice'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
        post:
            operationId: CreateExperiment
            x-contract-operation-id: ops.experiment.CreateExperiment
            x-object-id: ops.experiment
            x-actor: account
            x-application:
                kind: command
                facet: ExperimentCommandFacade
                method: createExperiment
                aggregateOwner: Experiment
            parameters:
                - name: Idempotency-Key
                  in: header
                  required: true
                  description: Stable business replay identity; distinct from X-Request-Id.
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateExperimentRequest'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateExperimentResult'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
    /control-plane/product/growth/overview:
        get:
            operationId: GetGrowthOverview
            x-contract-operation-id: ops.event_record.GetGrowthOverview
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordMetricQueryFacade
                method: getGrowthOverview
                reader: GrowthOverviewReader
                slice: GrowthOverview
            parameters:
                - name: days
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetGrowthOverviewResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
    /control-plane/product/projections/summary:
        get:
            operationId: GetProductProjectionSummary
            x-contract-operation-id: ops.event_record.GetProductProjectionSummary
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordControlPlaneQueryFacade
                method: getProductProjectionSummary
                reader: ProductProjectionSummaryReader
                slice: ProductProjectionSummary
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetProductProjectionSummaryResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
    /control-plane/product/triage/summary:
        get:
            operationId: GetProductTriageSummary
            x-contract-operation-id: ops.event_record.GetProductTriageSummary
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordControlPlaneQueryFacade
                method: getProductTriageSummary
                reader: ProductTriageSummaryReader
                slice: ProductTriageSummary
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetProductTriageSummaryResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /control-plane/product/workflows:
        get:
            operationId: ListProductWorkflows
            x-contract-operation-id: ops.event_record.ListProductWorkflows
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventRecordControlPlaneQueryFacade
                method: listProductWorkflows
                reader: ProductWorkflowReader
                slice: ProductWorkflowSlice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListProductWorkflowsResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
    /ops/events/drilldown:
        get:
            operationId: GetEventDrilldown
            x-contract-operation-id: ops.event_record.GetEventDrilldown
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventAnalyticsQueryFacade
                method: getDrilldown
                reader: EventDrilldownReader
                slice: EventDrilldownSlice
            parameters:
                - name: logType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: eventType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: pageName
                  in: query
                  required: false
                  schema:
                    type: string
                - name: appVersion
                  in: query
                  required: false
                  schema:
                    type: string
                - name: networkClass
                  in: query
                  required: false
                  schema:
                    type: string
                - name: result
                  in: query
                  required: false
                  schema:
                    type: string
                - name: errorCode
                  in: query
                  required: false
                  schema:
                    type: string
                - name: sessionId
                  in: query
                  required: false
                  schema:
                    type: string
                - name: from
                  in: query
                  required: false
                  schema:
                    type: string
                - name: to
                  in: query
                  required: false
                  schema:
                    type: string
                - name: limit
                  in: query
                  required: false
                  schema:
                    type: string
                - name: revealSession
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventDrilldownSlice'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /ops/events/rtc-media-qoe/summary:
        get:
            operationId: GetRtcMediaQoeSummary
            x-contract-operation-id: ops.event_record.GetRtcMediaQoeSummary
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: RtcMediaQoeQueryFacade
                method: get24HourSummary
                reader: RtcMediaQoeSummaryReader
                slice: RtcMediaQoeSummarySlice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RtcMediaQoeSummarySlice'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /ops/events/summary:
        get:
            operationId: GetEventSummary
            x-contract-operation-id: ops.event_record.GetEventSummary
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: EventAnalyticsQueryFacade
                method: getSummary
                reader: EventSummaryReader
                slice: EventSummarySlice
            parameters:
                - name: logType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: eventType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: pageName
                  in: query
                  required: false
                  schema:
                    type: string
                - name: appVersion
                  in: query
                  required: false
                  schema:
                    type: string
                - name: networkClass
                  in: query
                  required: false
                  schema:
                    type: string
                - name: result
                  in: query
                  required: false
                  schema:
                    type: string
                - name: errorCode
                  in: query
                  required: false
                  schema:
                    type: string
                - name: from
                  in: query
                  required: false
                  schema:
                    type: string
                - name: to
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetEventSummaryResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
    /ops/runtime-logs:
        post:
            operationId: ReportRuntimeLogBatch
            x-contract-operation-id: ops.event_record.ReportRuntimeLogBatch
            x-object-id: ops.event_record
            x-actor: persona_or_device
            x-application:
                kind: command
                facet: RuntimeLogCommandFacet
                method: appendAppRuntimeBatch
                appendSink: EventRecord
            parameters:
                - name: Idempotency-Key
                  in: header
                  required: true
                  description: Stable business replay identity; distinct from X-Request-Id.
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RuntimeLogBatchRequest'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventRecordBatchReceipt'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /ops/runtime-logs/drilldown:
        get:
            operationId: GetRuntimeLogDrilldown
            x-contract-operation-id: ops.event_record.GetRuntimeLogDrilldown
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: RuntimeLogQueryFacet
                method: getDrilldown
                reader: RuntimeLogDrilldownReader
                slice: RuntimeLogDrilldownSlice
            parameters:
                - name: signal
                  in: query
                  required: false
                  schema:
                    type: string
                - name: severity
                  in: query
                  required: false
                  schema:
                    type: string
                - name: errorCode
                  in: query
                  required: false
                  schema:
                    type: string
                - name: fingerprint
                  in: query
                  required: false
                  schema:
                    type: string
                - name: sourceType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: service
                  in: query
                  required: false
                  schema:
                    type: string
                - name: appVersion
                  in: query
                  required: false
                  schema:
                    type: string
                - name: actorHash
                  in: query
                  required: false
                  schema:
                    type: string
                - name: messageContains
                  in: query
                  required: false
                  schema:
                    type: string
                - name: from
                  in: query
                  required: false
                  schema:
                    type: string
                - name: to
                  in: query
                  required: false
                  schema:
                    type: string
                - name: limit
                  in: query
                  required: false
                  schema:
                    type: string
                - name: revealCorrelation
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetRuntimeLogDrilldownResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
    /ops/runtime-logs/summary:
        get:
            operationId: GetRuntimeLogSummary
            x-contract-operation-id: ops.event_record.GetRuntimeLogSummary
            x-object-id: ops.event_record
            x-actor: account
            x-application:
                kind: query
                facet: RuntimeLogQueryFacet
                method: getSummary
                reader: RuntimeLogSummaryReader
                slice: RuntimeLogSummarySlice
            parameters:
                - name: signal
                  in: query
                  required: false
                  schema:
                    type: string
                - name: severity
                  in: query
                  required: false
                  schema:
                    type: string
                - name: errorCode
                  in: query
                  required: false
                  schema:
                    type: string
                - name: fingerprint
                  in: query
                  required: false
                  schema:
                    type: string
                - name: sourceType
                  in: query
                  required: false
                  schema:
                    type: string
                - name: service
                  in: query
                  required: false
                  schema:
                    type: string
                - name: appVersion
                  in: query
                  required: false
                  schema:
                    type: string
                - name: from
                  in: query
                  required: false
                  schema:
                    type: string
                - name: to
                  in: query
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetRuntimeLogSummaryResponse'
                default:
                    description: Runtime failure
                    content:
                        application/json:
                            schema:
                                $ref: ../_shared/openapi_common.yaml#/components/schemas/ErrorResponse
            security:
                - bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            $ref: ../_shared/openapi_common.yaml#/components/securitySchemes/bearerAuth
    schemas:
...
        CreateExperimentRequest:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: CreateExperimentRequest
            x-contract-placeholder: true
        CreateExperimentResult:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: CreateExperimentResult
            x-contract-placeholder: true
...
        EventDrilldownSlice:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: EventDrilldownSlice
            x-contract-placeholder: true
        EventRecordBatchReceipt:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: EventRecordBatchReceipt
            x-contract-placeholder: true
...
        ExperimentCatalogSlice:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: ExperimentCatalogSlice
            x-contract-placeholder: true
...
        GetEventSummaryResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetEventSummaryResponse
            x-contract-placeholder: true
...
        GetGrowthOverviewResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetGrowthOverviewResponse
            x-contract-placeholder: true
        GetPageExperienceResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetPageExperienceResponse
            x-contract-placeholder: true
...
        GetProductProjectionSummaryResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetProductProjectionSummaryResponse
            x-contract-placeholder: true
        GetProductTriageSummaryResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetProductTriageSummaryResponse
            x-contract-placeholder: true
        GetRuntimeLogDrilldownResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetRuntimeLogDrilldownResponse
            x-contract-placeholder: true
        GetRuntimeLogSummaryResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: GetRuntimeLogSummaryResponse
            x-contract-placeholder: true
...
        ListProductApprovalsResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: ListProductApprovalsResponse
            x-contract-placeholder: true
        ListProductAuditsResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: ListProductAuditsResponse
            x-contract-placeholder: true
        ListProductWorkflowsResponse:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: ListProductWorkflowsResponse
            x-contract-placeholder: true
...
        RtcMediaQoeSummarySlice:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: RtcMediaQoeSummarySlice
            x-contract-placeholder: true
        RuntimeLogBatchRequest:
            type: object
            description: 字段由 fields/projection codegen 契约定义；当前 OpenAPI compiler 仅保留命名引用。
            x-contract-entity: RuntimeLogBatchRequest
            x-contract-placeholder: true
...